==> docs/articles/AutoML24Workshop_MScThesis/03_MSc_preprocessing.Rmd <==
---
title: "Masters Thesis forester: Preprocessing"
author: "Hubert Ruczyński"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    toc_collapsed: yes
    theme: lumen
    toc_depth: 3
    number_sections: yes
    code_folding: hide
    latex_engine: xelatex
---

```{css, echo=FALSE}
body .main-container {
  max-width: 1820px !important;
  width: 1820px !important;
}
body {
  max-width: 1820px !important;
  width: 1820px !important;
  font-family: Helvetica !important;
  font-size: 16pt !important;
}
h1,h2,h3,h4,h5,h6{
  font-size: 24pt !important;
}
```

# Downloads

The necessary downloads required for the forester package to work properly, if downloaded, the user can skip this part.

```{r eval = FALSE}
install.packages("devtools")
devtools::install_github("ModelOriented/forester")
devtools::install_url('https://github.com/catboost/catboost/releases/download/v1.2.2/catboost-R-Windows-1.2.2.tgz', INSTALL_opts = c("--no-multiarch", "--no-test-load"))
devtools::install_github('ricardo-bion/ggradar', dependencies = TRUE)
install.packages('tinytex')
tinytex::install_tinytex()
```

# Imports

Importing the necessary libraries.

```{r warning=FALSE, message=FALSE}
library(forester)
```

# Data summary

This short analysis presents the datasets used for the experiments, describing their basic characteristic as well as issues detected with `check_data()`.

```{r eval = FALSE}
MSc_binary_CC18      <- readRDS("MSc_binary_CC18.RData")
MSc_binary_CC18      <- MSc_binary_CC18[c(1, 2, 3, 4, 5, 19, 25, 26, 36, 37)]
MSc_multiclass_CC18  <- readRDS("MSc_multiclass_CC18.RData")
MSc_regression_bench <- readRDS("MSc_regression_bench.RData")
```

The results of this summary were created manually mostly based on the script `ablation_study_datasets_info.Rmd`. As seen below they are saved in the file named `data_issues_summary.csv`.

```{r eval = FALSE}
data_summary <- read.csv('data_issues_summary.csv', sep = ';')
rmarkdown::paged_table(data_summary)
```

# Single experiment

This section contains the code necessary for running the preprocessing for a single dataset.

## Removal settings

Firstly we define three removal patterns, as testing all possible combinations is unnecessary and incredibly time-consuming. We call those options: removal_min, removal_med, and removal_max. The first option runs a minimal preprocessing pipeline where we remove only observations that have no target value. The second one additionally removes duplicate, id-like, static, and sparse columns removing corrupted rows with too many missing values. The last option additionally includes the removal of highly correlated columns.

```{r eval = FALSE}
removal_min <- list(active_modules = c(
                      duplicate_cols = FALSE, id_like_cols = FALSE, static_cols = FALSE,
                      sparse_cols = FALSE, corrupt_rows = FALSE, correlated_cols = FALSE
                    ),
                    id_names                   = c(''),
                    static_threshold           = 1,
                    sparse_columns_threshold   = 1,
                    sparse_rows_threshold      = 1,
                    high_correlation_threshold = 1
                  )

removal_med <- list(active_modules = c(
                      duplicate_cols = TRUE, id_like_cols = TRUE, static_cols = TRUE,
                      sparse_cols = TRUE, corrupt_rows = TRUE, correlated_cols = FALSE
                    ),
                    id_names                   = c(''),
                    static_threshold           = 0.99,
                    sparse_columns_threshold   = 0.3,
                    sparse_rows_threshold      = 0.3,
                    high_correlation_threshold = 1
                  )

removal_max <- list(active_modules = c(
                      duplicate_cols = TRUE, id_like_cols = TRUE, static_cols = TRUE,
                      sparse_cols = TRUE, corrupt_rows = TRUE, correlated_cols = TRUE
                    ),
                    id_names                   = c(''),
                    static_threshold           = 0.99,
                    sparse_columns_threshold   = 0.3,
                    sparse_rows_threshold      = 0.3,
                    high_correlation_threshold = 0.7
                  )
```

## Parameters preparation

To run the preprocessing we prepare lists and vectors of parameters that will be used in the `custom_preprocessing()` function. We want to test various scenarios where:

-   RM - tests all removal strategies, where other modules are set to the most basic options: median-other for imputation, and none for feature selection. (3)
-   IMP - tests all imputation methods, where other modules are set to the most basic options: min for removals, and none for feature selection. (4)
-   FS - tests all feature selection methods, where other modules are set to the most basic options: min for removals, and median-other for imputation. (4)
-   RM + IMP - tests med and max removal options combined with all imputation methods. (8)
-   RM + FS - tests med and max removal options combined with all feature selection methods. (8)
-   IMP + FS - tests all imputation methods with MI and Boruta feature selection. (8)
-   RM + IMP + FS - tests med and max removal options combined with median-frequency, median-other and KNN imputation, and MI and Boruta feature selection. (12)

Finally we end up with 38 different sets preprocessing for a single dataset.

```{r eval = FALSE}
removals <- list(removal_min, removal_med, removal_max,              # RM
                 removal_min, removal_min, removal_min,              # Imp
                 removal_min, removal_min, removal_min, removal_min, # FS
                 removal_med, removal_med, removal_med,              # RM + Imp 
                 removal_max, removal_max, removal_max,              # RM + Imp
                 removal_med, removal_med, removal_med, removal_med, # RM + FS
                 removal_max, removal_max, removal_max, removal_max, # RM + FS
                 removal_min, removal_min, removal_min,              # Imp + FS
                 removal_min, removal_min, removal_min,              # Imp + FS 
                 removal_med, removal_med, removal_med, removal_med, # RM + Imp + FS 
                 removal_max, removal_max, removal_max, removal_max) # RM + Imp + FS 

imp_method <- c('median-other',     'median-other',     'median-other',                     # RM
                'median-frequency', 'knn',              'mice',                             # Imp
                'median-other',     'median-other',     'median-other',     'median-other', # FS
                'median-frequency', 'knn',              'mice',                             # RM + Imp 
                'median-frequency', 'knn',              'mice',                             # RM + Imp 
                'median-other',     'median-other',     'median-other',     'median-other', # RM + FS
                'median-other',     'median-other',     'median-other',     'median-other', # RM + FS
                'median-frequency', 'knn',              'mice',                             # Imp + FS 
                'median-frequency', 'knn',              'mice',                             # Imp + FS 
                'median-frequency',  'knn',             'median-frequency', 'knn',          # RM + Imp + FS 
                'median-frequency',  'knn',             'median-frequency', 'knn')          # RM + Imp + FS

fs_method <- c('none',   'none',   'none',             # RM
               'none',   'none',   'none',             # Imp
               'VI',     'MCFS',   'MI',     'BORUTA', # FS
               'none',   'none',   'none',   'none',   # RM + Imp 
               'none',   'none',   'none',   'none',   # RM + Imp 
               'VI',     'MCFS',   'MI',     'BORUTA', # RM + FS
               'VI',     'MCFS',   'MI',     'BORUTA', # RM + FS
               'MI',     'MI',     'MI',               # Imp + FS 
               'BORUTA', 'BORUTA', 'BORUTA',           # Imp + FS
               'MI',     'MI',     'BORUTA', 'BORUTA', # RM + Imp + FS
               'MI',     'MI',     'BORUTA', 'BORUTA') # RM + Imp + FS

rmv_names <- c('removal_min', 'removal_med', 'removal_max',                # RM
               'removal_min', 'removal_min', 'removal_min',                # Imp
               'removal_min', 'removal_min', 'removal_min', 'removal_min', # FS
               'removal_med', 'removal_med', 'removal_med',                # RM + Imp 
               'removal_max', 'removal_max', 'removal_max',                # RM + Imp
               'removal_med', 'removal_med', 'removal_med', 'removal_med', # RM + FS
               'removal_max', 'removal_max', 'removal_max', 'removal_max', # RM + FS
               'removal_min', 'removal_min', 'removal_min',                # Imp + FS 
               'removal_min', 'removal_min', 'removal_min',                # Imp + FS 
               'removal_med', 'removal_med', 'removal_med', 'removal_med', # RM + Imp + FS 
               'removal_max', 'removal_max', 'removal_max', 'removal_max') # RM + Imp + FS 
```

## Experiment function

This is the main experiment function working for a singular dataset. To make it work we need to provide the data, specify the target, provide a string name of the dataset, include previously defined parameter settings (`removals`, `imp_method`, `fs_method`, `rmv_names`), specify if the dataset is binary classification, and choose MI method (in the end it was `estevez` for binary classification, and `peng` for regression). The function can be run in silent or informative mode.

The function prepares particular datasets one by one, and the results are saved after into the sub-directories in order to enable stopping and resuming whole process. It creates a folder `preprocessing_data` in the working directory where the results are held. Inside we will find a sub-directory for each dataset where we hold 38 RData files with preprocessed data, as well as 2 helper RData files with `list_times`, and `names_times`. When the process is finished we will also obtain another RData file with list of all preprocessed versions of the dataset in the main `preprocessing_data` directory. Full analysis consists of 25 sub-directories and final RData files.

```{r eval = FALSE}
preprocessing_experiment <- function(data, y, removals, imp_method, fs_method, 
                                     dataset_name, rmv_names, task = 'binary', verbose = TRUE,
                                     mi_method = 'estevez') {
  list_times  <- c()
  names_times <- c()
  prep_data   <- list()
  dir.create('MSc_preprocessing_data', showWarnings = FALSE)
  
  if (task == 'binary') {
    dir.create(paste0('MSc_preprocessing_data/binary_exp_', dataset_name), showWarnings = FALSE)
    tryCatch({
      suppressWarnings(list_times  <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/binary_exp_', 
                                                     dataset_name, '/list_times.RData')))
      suppressWarnings(names_times <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/binary_exp_', 
                                                     dataset_name, '/names_times.RData')))
      print('Loaded times list.')
    }, error = function(cond) {
      print('No times list.')
    })
  } else if (task == 'regression') {
    dir.create(paste0('MSc_preprocessing_data/regression_exp_', dataset_name), showWarnings = FALSE)
    tryCatch({
      suppressWarnings(list_times  <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/regression_exp_', 
                                                     dataset_name, '/list_times.RData')))
      suppressWarnings(names_times <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/regression_exp_', 
                                                     dataset_name, '/names_times.RData')))
      print('Loaded times list.')
    }, error = function(cond) {
      print('No times list.')
    })
  } else if (task == 'multiclass') {
    dir.create(paste0('MSc_preprocessing_data/multiclass_exp_', dataset_name), showWarnings = FALSE)
    tryCatch({
      suppressWarnings(list_times  <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/multiclass_exp_', 
                                                     dataset_name, '/list_times.RData')))
      suppressWarnings(names_times <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/multiclass_exp_', 
                                                     dataset_name, '/names_times.RData')))
      print('Loaded times list.')
    }, error = function(cond) {
      print('No times list.')
    })
  }
  
  for (i in 1:38) {
    verbose_cat('\n Iteration:', i, 'Removal:', rmv_names[i], 'Imputation:', 
                imp_method[i], 'Feature Selection:', fs_method[i], '\n', verbose = verbose)


    if (!file.exists(paste0('MSc_preprocessing_data/binary_exp_', dataset_name, '/', i, '.RData')) &&
        !file.exists(paste0('MSc_preprocessing_data/regression_exp_', dataset_name, '/', i,  '.RData')) &&
        !file.exists(paste0('MSc_preprocessing_data/multiclass_exp_', dataset_name, '/', i,  '.RData'))) {
        start          <- as.numeric(Sys.time())
        names_times[i] <- paste0('Training: ', dataset_name, ' Removal: ', rmv_names[i], 
                                 ' Imputation: ', imp_method[i], ' Feature Selection: ', fs_method[i])
        script_wd      <- getwd()
        prep <- custom_preprocessing(data = data,
                                     y    = y,
                                     na_indicators         = c(''),
                                     removal_parameters    = removals[[i]],
                                     imputation_parameters = list(
                                       imputation_method = imp_method[[i]],
                                       k = 10,
                                       m = 5
                                     ),
                                     feature_selection_parameters = list(
                                       feature_selection_method = fs_method[[i]],
                                       max_features       = 'default',
                                       nperm              = 1,
                                       cutoffPermutations = 20,
                                       threadsNumber      = NULL,
                                       method             = mi_method
                                     ),
                                     verbose = TRUE)
        setwd(script_wd)
        stop           <- as.numeric(Sys.time())
        list_times[i]  <- round(stop - start, 1)
        prep_data[[i]] <- prep
        
        if (task == 'binary') {
          saveRDS(prep, paste0(getwd(), '/MSc_preprocessing_data/binary_exp_', dataset_name, '/', i, '.RData'))
          saveRDS(list_times,  paste0(getwd(), '/MSc_preprocessing_data/binary_exp_', dataset_name, '/list_times.RData'))
          saveRDS(names_times, paste0(getwd(), '/MSc_preprocessing_data/binary_exp_', dataset_name, '/names_times.RData'))
        } else if (task == 'regression') {
          saveRDS(prep, paste0(getwd(), '/MSc_preprocessing_data/regression_exp_', dataset_name, '/', i, '.RData'))
          saveRDS(list_times,  paste0(getwd(), '/MSc_preprocessing_data/regression_exp_', dataset_name, '/list_times.RData'))
          saveRDS(names_times, paste0(getwd(), '/MSc_preprocessing_data/regression_exp_', dataset_name, '/names_times.RData'))
        } else if (task == 'multiclass') {
          saveRDS(prep, paste0(getwd(), '/MSc_preprocessing_data/multiclass_exp_', dataset_name, '/', i, '.RData'))
          saveRDS(list_times,  paste0(getwd(), '/MSc_preprocessing_data/multiclass_exp_', dataset_name, '/list_times.RData'))
          saveRDS(names_times, paste0(getwd(), '/MSc_preprocessing_data/multiclass_exp_', dataset_name, '/names_times.RData'))
        }
    } else if (task == 'binary') {
      prep_data[[i]] <- readRDS(paste0('MSc_preprocessing_data/binary_exp_', dataset_name, '/', i, '.RData'))
    } else if (task == 'regression') {
      prep_data[[i]] <- readRDS(paste0('MSc_preprocessing_data/regression_exp_', dataset_name, '/', i, '.RData'))
    } else if (task == 'multiclass') {
      prep_data[[i]] <- readRDS(paste0('MSc_preprocessing_data/multiclass_exp_', dataset_name, '/', i, '.RData'))
    }

  }
  names(prep_data) <- names_times
  time_df          <- data.frame(name = names_times, duration = list_times)
  
  outcome <- list(
    time_df   = time_df,
    prep_data = prep_data
  )
  
  if (task == 'binary') {
    saveRDS(outcome, paste0('MSc_preprocessing_data/binary_exp_', dataset_name, '.RData'))
  } else if (task == 'regression') {
    saveRDS(outcome, paste0('MSc_preprocessing_data/regression_exp_', dataset_name, '.RData'))
  } else if (task == 'multiclass') {
    saveRDS(outcome, paste0('MSc_preprocessing_data/multiclass_exp_', dataset_name, '.RData'))
  }
}
```

# All datasets experiments

In this section we present the function conducting multiple experiments for different datasets.

## Parameters preparation

Firstly we prepare lists of additional parameters describing the datasets.

```{r eval = FALSE}
targets_binary     <- c('class', 'Class', 'class', 'class', 'class', 'Class', 'Class', 'Class', 'class', 'Class')
targets_regression <- c('rej', 'quality', 'y', 'y', 'foo', 'y', 'Goal', 'Goal', 'y')
targets_multiclass <- c('class', 'class', 'class', 'class', 'class', 'class', 'class', 'class', 'class')
names_binary       <- c('kr-vs-kp', 'breast-w', 'credit-approval', 'credit-g', 'diabetes', 
                        'phoneme', 'banknote-authentication', 'blood-transfusion-service-center',
                        'credit-g-mod', 'phoneme-mod')
names_regression   <- c('bank32nh', 'wine_quality', 'Mercedes_Benz_Greener_Manufacturing', 
                        'kin8nm', 'pol', '2dplanes', 'elevators', 'elevators-mod', 'kin8nm-mod')
names_multiclass   <- c("balance-scale", "mfeat-karhunen", "mfeat-zernike", "satimage", "car", "segment", "dna", "satimage-mod", "car-mod" )
```

## Multiple experiments function

The function is similar to the previous one, and it helps with executing the preprocessing of multiple datasets at once.

```{r eval = FALSE}
multiple_experiments <- function(datasets, targets, removals, imp_method, fs_method, 
                                 dataset_names, rmv_names, task = 'binary', verbose = 'part',
                                 mi_method = 'estevez') {
  
  if (verbose == 'part') {
    text_verbose <- TRUE
    exp_verbose  <- FALSE
  } else if (verbose == 'all') {
    text_verbose <- TRUE
    exp_verbose  <- TRUE
  } else if (verbose == 'none') {
    text_verbose <- FALSE
    exp_verbose  <- FALSE
  }
  
  for (i in 1:length(targets)) {
    if (!file.exists(paste0('MSc_preprocessing_data/binary_exp_', dataset_names[i], '.RData')) &&
        !file.exists(paste0('MSc_preprocessing_data/regression_exp_', dataset_names[i], '.RData')) &&
        !file.exists(paste0('MSc_preprocessing_data/multiclass_exp_', dataset_names[i], '/', i,  '.RData'))) {
      
      verbose_cat('The file for the', i, 'dataset, called', dataset_names[i], 
                  'does not exist, proceeding with preprocessing.\n', verbose = text_verbose)
      
      preprocessing_experiment(data         = datasets[[i]], 
                               y            = targets[i], 
                               removals     = removals, 
                               imp_method   = imp_method, 
                               fs_method    = fs_method,
                               dataset_name = dataset_names[i],
                               rmv_names    = rmv_names,
                               task         = task,
                               verbose      = exp_verbose,
                               mi_method    = mi_method)
    } else {
      verbose_cat('The file for the', i, 'dataset, called', dataset_names[i], 
                  'exists, skipping the preprocessing.\n', verbose = text_verbose)
    }
  }
}
```

# Preprocessing execution

The code executing the preprocessing for both binary classification and regression tasks.

## Binary classification

```{r eval = FALSE}
idx <- 1:10
multiple_experiments(datasets      = MSc_binary_CC18[idx], 
                     targets       = targets_binary[idx], 
                     removals      = removals, 
                     imp_method    = imp_method, 
                     fs_method     = fs_method, 
                     dataset_names = names_binary[idx], 
                     rmv_names     = rmv_names, 
                     task          = 'binary', 
                     verbose       = 'part',
                     mi_method     = 'estevez')
```

## Regression

```{r eval = FALSE}
idx <- 1:9
multiple_experiments(datasets      = MSc_regression_bench[idx], 
                     targets       = targets_regression[idx], 
                     removals      = removals, 
                     imp_method    = imp_method, 
                     fs_method     = fs_method, 
                     dataset_names = names_regression[idx], 
                     rmv_names     = rmv_names, 
                     task          = 'regression', 
                     verbose       = 'part',
                     mi_method     = 'peng')
```

## Multiclass

```{r eval = FALSE}
idx <- 1:9
multiple_experiments(datasets      = MSc_multiclass_CC18[idx], 
                     targets       = targets_multiclass[idx], 
                     removals      = removals, 
                     imp_method    = imp_method, 
                     fs_method     = fs_method, 
                     dataset_names = names_multiclass[idx], 
                     rmv_names     = rmv_names, 
                     task          = 'multiclass', 
                     verbose       = 'part',
                     mi_method     = 'peng')
```

# Short time analysis

```{r}
times <- list()
full_times <- c()
binary_names       <- paste0('binary_exp_', names_binary)
regression_names   <- paste0('regression_exp_', names_regression)
multiclass_names   <- paste0('multiclass_exp_', names_multiclass)
all_names          <- c(binary_names, regression_names, multiclass_names)
for (i in 1:25) {
  times[[i]] <- readRDS(paste0(getwd(), '/MSc_preprocessing_data/', all_names[i], '/list_times.RData'))
  full_times <- c(full_times, sum(times[[i]]))
}


```

```{r}
cat('Full preprocessing times in seconds:\n', full_times, '\n\n')
cat('Full preprocessing times in minutes:\n', round(full_times/60, 2), '\n\n')
cat('Full preprocessing times in hours:\n', round(full_times/3600, 2), '\n\n')
```

```{r}
files <- list.files('MSc_preprocessing_data', pattern = 'RData')
data  <- list()
for (i in 1:length(files)) {
  data[[i]] <- readRDS(paste0('MSc_preprocessing_data/', files[i]))
}
```

```{r}
dataset           <- c()
removal           <- c()
imputation        <- c()
feature_selection <- c()
duration          <- c()
task_type         <- c()
for (i in 1:25) {
  for (j in 1:38) {
    stringsplt        <- strsplit(data[[i]]$time_df$name[[j]], ':')[[1]]
    dataset           <- c(dataset,    substr(stringsplt[2], 2, nchar(stringsplt[2]) - 7))
    removal           <- c(removal,    substr(stringsplt[3], 1, nchar(stringsplt[3]) - 10))
    imputation        <- c(imputation, substr(stringsplt[4], 1, nchar(stringsplt[4]) - 17))
    feature_selection <- c(feature_selection, stringsplt[5])
    duration          <- c(duration, data[[i]]$time_df$duration[j])
    if (i <= 10) {
      task_type <- c(task_type, 'binary')
    } else if (i > 18) {
      task_type <- c(task_type, 'regression')
    } else {
      task_type <- c(task_type, 'multiclass')
    }
  }
}
duration_df <- data.frame(Dataset = dataset, Removal = removal, Imputation = imputation, 
                          Feature_selection = feature_selection, Task_type = task_type, Duration = duration)
rmarkdown::paged_table(duration_df)
```
```{r}
sum(duration_df$Duration[duration_df$Task_type == 'binary'])/3600
sum(duration_df$Duration[duration_df$Task_type == 'regression'])/3600
sum(duration_df$Duration[duration_df$Task_type == 'multiclass'])/3600
```

```{r}
dir.create('MSc_processed_results', showWarnings = FALSE)
#saveRDS(duration_df, 'MSc_processed_results/preprocessing_duration.RData')
```
