==> docs/articles/AutoML24Workshop_MScThesis/06_MSc_preliminary_results_analysis.Rmd <==
---
title: "Masters Thesis forester: Preliminary results analysis"
author: "Hubert Ruczyński"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    toc_collapsed: yes
    theme: lumen
    toc_depth: 3
    number_sections: yes
    code_folding: hide
    latex_engine: xelatex
---

```{css, echo=FALSE}
body .main-container {
  max-width: 1820px !important;
  width: 1820px !important;
}
body {
  max-width: 1820px !important;
  width: 1820px !important;
  font-family: Helvetica !important;
  font-size: 16pt !important;
}
h1,h2,h3,h4,h5,h6{
  font-size: 24pt !important;
}
```

# Important note

In this notebook you will find preliminary results of our study, however none of these plots are used in the Master's Thesis. This supplementary material is a form of additional knowledge, if some readers do not like aggregations. To see the final results, please refer to the notebooks 07 and 08.

# Imports and settings

```{r, warning=FALSE, message=FALSE}
library(ggplot2)
library(patchwork)
library(scales)
library(dplyr)
library(forcats)
```


# Data import

```{r}
duration_train_df      <- readRDS('MSc_processed_results/training_duration.RData')
duration_preprocessing <- readRDS('MSc_processed_results/preprocessing_duration.RData')
training_summary_table <- readRDS('MSc_processed_results/training_summary_table.RData')
testing_summary_table  <- readRDS('MSc_processed_results/testing_summary_table.RData')
validation_summary_table <- readRDS('MSc_processed_results/validation_summary_table.RData')
```

# Time analysis

An important aspect of our analysis is the time complexity of different approaches, as extended preprocessing module leads to more time consuming computations, which could be spent for example on training the models. On the other hand, thorough preparation step might result in removing lots of unnecessary columns, so the model should be able to learn faster. Despite the absolute preprocessing time, another important aspect is the relative duration to training time. Ex. if the training takes 1000 seconds than preprocessing lasting 100 is not so much as in the case when training takes 100 seconds. We will work on slightly modified data frame presented below.

```{r}
duration_df                                 <- duration_train_df
full_duration                               <- duration_preprocessing$Duration + duration_df$Duration
duration_df$Preprocessing_duration          <- duration_preprocessing$Duration
duration_df$Preprocessing_duration_fraction <- round(duration_df$Preprocessing_duration / full_duration, 3)
duration_df$Full_duration                   <- full_duration
rmarkdown::paged_table(duration_df)
```

## Training time

```{r, echo=FALSE}
column_fractions <- c()
max_fields_num   <- c()
task_type        <- c()
Columns          <- c()
datasets         <- unique(training_summary_table$Dataset)
for (i in 1:length(unique(training_summary_table$Dataset))) {
  cols <- training_summary_table[training_summary_table$Dataset == datasets[i], 'Columns']
  rows <- training_summary_table[training_summary_table$Dataset == datasets[i], 'Rows']
  Columns          <- c(Columns, max(cols))
  column_fractions <- c(column_fractions, round(min(cols) / max(cols), 2))
  max_fields_num   <- c(max_fields_num, max(rows) * max(cols))
  if (i <= 10) {
    task_type <- c(task_type, 'binary')
  } else if (i > 18) {
    task_type <- c(task_type, 'regression')
  } else {
    task_type <- c(task_type, 'multiclass')
  }
}
left_columns <- data.frame(Dataset = datasets, Column_fraction = column_fractions, Columns = Columns, 
                           Max_fields_number = max_fields_num, Task_type = task_type)
```

```{r}
left_columns_binary <- left_columns[left_columns$Task_type == 'binary', ]
#left_columns_binary$Dataset <- droplevels(left_columns_binary$Dataset)
left_columns_binary$Dataset <- fct_reorder(left_columns_binary$Dataset, left_columns_binary$Max_fields_number)

left_columns_multiclass <- left_columns[left_columns$Task_type == 'multiclass', ]
#left_columns_multiclass$Dataset <- droplevels(left_columns_multiclass$Dataset)
left_columns_multiclass$Dataset <- fct_reorder(left_columns_multiclass$Dataset, left_columns_multiclass$Max_fields_number)

left_columns_regression <- left_columns[left_columns$Task_type == 'regression', ]
#left_columns_regression$Dataset <- droplevels(left_columns_regression$Dataset)
left_columns_regression$Dataset <- fct_reorder(left_columns_regression$Dataset, left_columns_regression$Max_fields_number)

left_columns <- rbind(left_columns_binary, left_columns_multiclass, left_columns_regression)
```

```{r}
paper_theme <- function() {
  theme_minimal() +
  theme(plot.title    = element_text(colour = 'black', size = 20),
        plot.subtitle = element_text(colour = 'black', size = 16),
        axis.title.x  = element_text(colour = 'black', size = 14),
        axis.title.y  = element_text(colour = 'black', size = 14),
        axis.text.y   = element_text(colour = "black", size = 12),
        axis.text.x   = element_text(colour = "black", size = 12),
        strip.background = element_rect(fill = "white", color = "white"),
        strip.text = element_text(size = 6 ),
        strip.text.y.right = element_text(angle = 0),
        legend.title = element_text(colour = 'black', size = 14),
        legend.text  = element_text(colour = "black", size = 12))
}
```


```{r fig.height=8, fig.width=18, echo=FALSE}
a <- ggplot(data = left_columns, aes(color = Task_type, fill = Task_type)) + 
  geom_segment(aes(x = Columns * column_fractions, xend = Columns, y = Dataset, yend = Dataset)) +
  geom_point(aes(x = Columns * column_fractions, y = Dataset), size = 3) +
  geom_point(aes(x = Columns, y = Dataset), size = 3) +
  labs(title = 'Columns range',
       subtitle = 'full columns vs maximal reduction',
       x = 'Number of columns',
       y = '',
       color = 'Task type',
       fill  = 'Task type') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE)  +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.text.y   = element_blank(),
        legend.position = "none")

b <- ggplot(data = duration_df, aes(x = Duration, y = Dataset, color = Task_type, fill = Task_type)) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Training time',
       subtitle = 'for different ML tasks',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Task type',
       fill  = 'Task type') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE)  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

(b | a) + plot_layout(widths = c(3, 1))
```

## Preprocessing time

```{r fig.height=8, fig.width=18, echo=FALSE}
c <- ggplot(data = left_columns, aes(x = Max_fields_number, y = Dataset, color = Task_type, fill = Task_type)) + 
  geom_col(alpha = 0.5) + 
  labs(title = 'Number of initial fields',
       subtitle = '',
       x = 'Number of fields',
       y = '',
       color = 'Task_type',
       fill  = 'Task_type') +
  scale_x_continuous(trans = log2_trans(), 
                     breaks = trans_breaks('log2', function(x) 2^x),
                     labels = trans_format('log2', math_format(2^.x))) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.text.y   = element_blank(),
        legend.position = "none")

d <- ggplot(data = duration_df, aes(x = Preprocessing_duration, y = Dataset, color = Task_type, fill = Task_type)) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time',
       subtitle = 'for different ML tasks',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Task type',
       fill  = 'Task type') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")
(d | a | c) + plot_layout(widths = c(3, 1, 1))
```

## Combined time

```{r fig.height=8, fig.width=18, echo=FALSE}
e <- ggplot(data = duration_df, aes(x = Full_duration, y = Dataset, color = Task_type, fill = Task_type)) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Combined preprocessing and training time',
       subtitle = 'for different ML tasks',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Task type',
       fill  = 'Task type') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

(e | a | c) + plot_layout(widths = c(3, 1, 1))
```

```{r fig.height=8, fig.width=18, echo=FALSE}
f <- ggplot(data = duration_df, aes(x = Preprocessing_duration_fraction, y = Dataset, color = Task_type, fill = Task_type)) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time fraction',
       subtitle = ' in comparison to full process, for different ML tasks',
       x = 'Fraction of preprocessing time',
       y = 'Dataset',
       color = 'Task type',
       fill  = 'Task type') +
  xlim(0, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

(f | a | c) + plot_layout(widths = c(3, 1, 1))
```

## Preprocessing components analysis

### Feature selection impact

```{r fig.height=8, fig.width=18, echo=FALSE}
bool_fs <- duration_preprocessing
bool_fs[bool_fs$Feature_selection != ' none', 'Feature_selection'] <- 'yes'

g <- ggplot(data = bool_fs, aes(x = Duration, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  theme_minimal() + 
  labs(title = 'Preprocessing time',
       subtitle = 'for different ML tasks, divided by presence of feature selection',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

g
```

### No feature selection removal strategies


```{r fig.height=8, fig.width=18, echo=FALSE}
no_fs <- duration_preprocessing[duration_preprocessing$Feature_selection == ' none', ]

h <- ggplot(data = no_fs, aes(x = Duration, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time',
       subtitle = 'for different ML tasks, divided by removal strategy',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

h
```

### No feature selection Imputation methods

```{r fig.height=8, fig.width=18, echo=FALSE}
no_fs_imp <- no_fs[no_fs$Dataset %in% c('breast-w ', 'credit-approval ', "credit-g-mod ", "phoneme-mod ", "car-mod ", "satimage-mod ", "elevators-mod ", "kin8nm-mod "), ]
i <- ggplot(data = no_fs_imp, aes(x = Duration, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time',
       subtitle = 'for different ML tasks, divided by imputation strategy',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Imputation strategy',
       fill  = 'Imputation strategy') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

j <- ggplot(data = no_fs, aes(x = Duration, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time comparison with forester',
       subtitle = 'for different ML tasks, divided by imputation strategy',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Imputation strategy',
       fill  = 'Imputation strategy') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

i
```

```{r fig.height=10, fig.width=18, echo=FALSE}
j
```

### Different feature selection methods

```{r fig.height=8, fig.width=18, echo=FALSE}
only_fs       <- duration_preprocessing[duration_preprocessing$Feature_selection != ' none', ]
only_fs_niche <- only_fs[only_fs$Feature_selection %in% c(' MI', ' MCFS'), ]
only_fs_top   <- only_fs[only_fs$Feature_selection %in% c(' VI', ' BORUTA'), ]

k <- ggplot(data = only_fs, aes(x = Duration, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Preprocessing time',
       subtitle = 'for different ML tasks, divided by feature selection method',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x), limits = c(NA, 4100)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")
k
```

```{r fig.height=8, fig.width=18, echo=FALSE}
datasets <- unique(only_fs$Dataset)
VI       <- c()
MCFS     <- c()
MI       <- c()
BORUTA   <- c()

for (i in unique(only_fs$Dataset)) {
  ds     <- only_fs[only_fs$Dataset == i, ]
  VI     <- c(VI,     median(ds[ds$Feature_selection == ' VI', 'Duration']))
  MCFS   <- c(MCFS,   median(ds[ds$Feature_selection == ' MCFS', 'Duration']))
  MI     <- c(MI,     median(ds[ds$Feature_selection == ' MI', 'Duration']))
  BORUTA <- c(BORUTA, median(ds[ds$Feature_selection == ' BORUTA', 'Duration']))
}

median_fs      <- data.frame(Dataset = datasets, VI = VI, MCFS = MCFS, BORUTA = BORUTA, MI = MI)
long_median_fs <- reshape(median_fs, varying = c('MI' ,'VI', 'MCFS', 'BORUTA'), v.names = c('Duration'), 
                          times = c('MI' ,'VI', 'MCFS', 'BORUTA'), direction = 'long')
long_median_fs <- long_median_fs[, 1:3]

rownames(long_median_fs) <- NULL
colnames(long_median_fs) <- c('Dataset', 'Method', 'Duration')

l <- ggplot(data = long_median_fs, aes(x = Duration, y = Dataset, color = factor(Method), fill = factor(Method))) + 
  geom_point(size = 5, alpha = 0.5) + 
  labs(title = 'Preprocessing median time',
       subtitle = 'for different ML tasks, divided by feature selection method',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")
l
```

The visualization above clearly indicates that in the forester package we can witness the division between slow and fast feature selection methods, where MCFS and MI are in the first group, whereas, VI and Borutain the second one. In order to analyse them thoroughly let's create two subplots that separate those two.

```{r fig.height=8, fig.width=18, echo=FALSE}
long_median_fs_slow <- long_median_fs[long_median_fs$Method %in% c('VI', 'BORUTA'), ]
long_median_fs_fast <- long_median_fs[long_median_fs$Method %in% c('MCFS', 'MI'), ]

m <- ggplot(data = long_median_fs_slow, aes(x = Duration, y = Dataset, color = factor(Method), fill = factor(Method))) + 
  geom_point(size = 5, alpha = 0.5) + 
  labs(title = 'Preprocessing median time',
       subtitle = 'for slow feature selection methods',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom")

n <- ggplot(data = long_median_fs_fast, aes(x = Duration, y = Dataset, color = factor(Method), fill = factor(Method))) + 
  geom_point(size = 5, alpha = 0.5) + 
  labs(title = 'Preprocessing median time',
       subtitle = 'for fast feature selection methods',
       x = 'Duration [s]',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  scale_x_continuous(trans = log2_trans(), breaks = trans_breaks('log2', function(x) 2^x)) +
  annotation_logticks(base = 2, scaled = TRUE) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.text.y   = element_blank(),
        axis.title.y  = element_blank(),
        legend.position = "bottom")
m | n
```

This time we can easily distinguish which preprocessing methods are faster and slower among considered pairs. In the case of less time-demanding ones presented on the right plot, every time MI method is faster than BORUTA, and in some cases the differences are significant as the cane reach up to 16 times difference. For the slow methods it is not so clear which one is more demanding, as sometimes VI is faster and sometimes MCFS. We could say that the slowest algorithm is the VI method, as there are 5 datasets where MCFS is incredibly fast, whereas the VI is much slower then.

Summing up, the order from fastest to slowest feature selection method is: MI, BORUTA, MCFS, VI.


# Performance

Now, let's analyse the performance of the models obtained in our experiment.

```{r, echo=FALSE}
all_engines               <- validation_summary_table[validation_summary_table$Engine == 'all', ]

all_engines_bin           <- all_engines[all_engines$Task_type == 'binary', ]
all_engines_mcl           <- all_engines[all_engines$Task_type == 'multiclass', ]
all_engines_reg           <- all_engines[all_engines$Task_type == 'regression', ]

all_engines_bin_baselines <- all_engines_bin[which(all_engines_bin$Removal =='removal_min' & 
                                                   all_engines_bin$Imputation =='median-other' & 
                                                   all_engines_bin$Feature_selection =='none'), ]
all_engines_mcl_baselines <- all_engines_mcl[which(all_engines_mcl$Removal =='removal_min' & 
                                                   all_engines_mcl$Imputation =='median-other' & 
                                                   all_engines_mcl$Feature_selection =='none'), ]
all_engines_reg_baselines <- all_engines_reg[which(all_engines_reg$Removal =='removal_min' & 
                                                   all_engines_reg$Imputation =='median-other' & 
                                                   all_engines_reg$Feature_selection =='none'), ]
```

## Comparison to baseline preprocessing

### Binary classification

```{r fig.height=12, fig.width=18, echo=FALSE, warning=FALSE}
o <- ggplot(data = all_engines_bin, aes(x = Max, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_bin_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Max, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Max metrics values',
       subtitle = 'for binary classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.35, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.title.x  = element_blank(),
        axis.title.y  = element_blank(),
        legend.position = "none")

p <- ggplot(data = all_engines_bin, aes(x = Mean, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) + 
  geom_point(data = all_engines_bin_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Mean, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Mean metrics values',
       subtitle = 'for binary classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.35, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(plot.subtitle   = element_blank(),
        axis.title.x    = element_blank(),
        legend.position = "none")

r <- ggplot(data = all_engines_bin, aes(x = Median, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) + 
  geom_point(data = all_engines_bin_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Median, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Median metrics values',
       subtitle = 'for binary classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.35, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(plot.subtitle   = element_blank(),
        axis.title.y    = element_blank(),
        legend.position = "bottom")

o / p / r
```


### Multiclass classification

```{r fig.height=12, fig.width=18, echo=FALSE, warning=FALSE}
o <- ggplot(data = all_engines_mcl, aes(x = Max, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_mcl_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Max, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Max metrics values',
       subtitle = 'for multiclass classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.25, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.title.y  = element_blank(),
        axis.title.x  = element_blank(),
        legend.position = "none")

p <- ggplot(data = all_engines_mcl, aes(x = Mean, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) + 
  geom_point(data = all_engines_mcl_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Mean, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Mean metrics values',
       subtitle = 'for multiclass classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.25, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(plot.subtitle   = element_blank(),
        axis.title.x    = element_blank(),
        legend.position = "none")

r <- ggplot(data = all_engines_mcl, aes(x = Median, y = Dataset, color = factor(Metric), fill = factor(Metric))) + 
  geom_boxplot(alpha = 0.5) + 
  geom_point(data = all_engines_mcl_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Median, y = Dataset, color = factor(Metric), fill = factor(Metric))) +
  labs(title = 'Median metrics values',
       subtitle = 'for multiclass classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Metric',
       fill  = 'Metric') +
  xlim(0.25, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(plot.subtitle   = element_blank(),
        axis.title.y    = element_blank(),
        legend.position = "bottom")

o / p / r
```

### Regression

```{r echo=FALSE}
all_engines_reg_min_med <- all_engines_reg[, c(1, 2, 3, 4, 5, 13, 16, 17)]
median                  <- all_engines_reg_min_med[, 1:7]
names(median)           <- c(names(median)[1:6], 'Value')
min                     <- all_engines_reg_min_med[, c(1:6, 8)]
names(min)              <- c(names(min)[1:6], 'Value')
all_engines_reg_min_med <- rbind(median, min)
all_engines_reg_min_med$Aggregation <- rep(c('Median', 'Min'), each = nrow(all_engines_reg))
```

```{r echo=FALSE}
all_engines_reg_baselines_min_med <- all_engines_reg_baselines[, c(1, 2, 3, 4, 5, 13, 16, 17)]
median                            <- all_engines_reg_baselines_min_med[, 1:7]
names(median)                     <- c(names(median)[1:6], 'Value')
min                               <- all_engines_reg_baselines_min_med[, c(1:6, 8)]
names(min)                        <- c(names(min)[1:6], 'Value')
all_engines_reg_baselines_min_med <- rbind(median, min)
all_engines_reg_baselines_min_med$Aggregation <- rep(c('Median', 'Min'), each = nrow(all_engines_reg_baselines))
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
metric <- 'mse'
s <- ggplot(data = all_engines_reg_min_med[which(all_engines_reg_min_med$Metric == metric), ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated MSE values (Magnified)',
       subtitle = 'for regression tasks, X-mark stands for baseline preprocessing',
       x        = 'Value',
       y        = 'Dataset',
       color    = 'Aggregation',
       fill     = 'Aggregation') +
  coord_cartesian(xlim = c(0, 2)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

t <- ggplot(data = all_engines_reg_min_med[all_engines_reg_min_med$Metric == metric, ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated MSE values (All)',
       x        = 'Value',) +
  coord_cartesian(xlim = c(0, NA)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank(),
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank())

metric <- 'mae'
u <- ggplot(data = all_engines_reg_min_med[which(all_engines_reg_min_med$Metric == metric), ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated MAE values (Magnified)',
       x        = 'Value',
       y        = 'Dataset',
       color    = 'Aggregation',
       fill     = 'Aggregation') +
  coord_cartesian(xlim = c(0, 2)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank())

v <- ggplot(data = all_engines_reg_min_med[all_engines_reg_min_med$Metric == metric, ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated MAE values (All)',
       x        = 'Value',) +
  coord_cartesian(xlim = c(0, NA)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank(),
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank())

metric <- 'rmse'
w <- ggplot(data = all_engines_reg_min_med[which(all_engines_reg_min_med$Metric == metric), ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated RMSE values (Magnified)',
       x        = 'Value',
       y        = 'Dataset',
       color    = 'Aggregation',
       fill     = 'Aggregation') +
  coord_cartesian(xlim = c(0, 2))  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank())

x <- ggplot(data = all_engines_reg_min_med[all_engines_reg_min_med$Metric == metric, ], 
            aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == metric), ], 
             size = 3, shape = 4,position = position_jitterdodge(), 
             aes(x = Value, y = Dataset, color = factor(Aggregation), fill = factor(Aggregation))) +
  labs(title    = 'Aggregated RMSE values (All)',
       x        = 'Value',) +
  coord_cartesian(xlim = c(0, NA)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank())

(s | t) / (u | v) / (w | x)
```

## Feature Selection Impact

### Binary classification

```{r echo=FALSE}
all_engines_bin_fs <- all_engines_bin
all_engines_bin_fs <- all_engines_bin_fs[all_engines_bin_fs$Metric == 'accuracy', ]
all_engines_bin_fs$Feature_selection <- ifelse(all_engines_bin_fs$Feature_selection != 'none', 'Yes', 'No')

```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
a1 <- ggplot(data = all_engines_bin_fs, aes(x = Max, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, wheter FS methods were used',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.7, 1)  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x   = element_blank())

b1 <- ggplot(data = all_engines_bin_fs, aes(x = Mean, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Mean Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.7, 1)  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank())

c1 <- ggplot(data = all_engines_bin_fs, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.7, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank())

a1 / c1

```

### Multiclass classification

```{r echo=FALSE}
all_engines_mcl_fs <- all_engines_mcl
all_engines_mcl_fs <- all_engines_mcl_fs[all_engines_mcl_fs$Metric == 'accuracy', ]
all_engines_mcl_fs$Feature_selection <- ifelse(all_engines_mcl_fs$Feature_selection != 'none', 'Yes', 'No')

```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
a1 <- ggplot(data = all_engines_mcl_fs, aes(x = Max, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks, wheter FS methods were used',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.5, 1)  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x   = element_blank())

b1 <- ggplot(data = all_engines_mcl_fs, aes(x = Mean, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Mean Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.5, 1)  + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank())

c1 <- ggplot(data = all_engines_mcl_fs, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) + 
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  xlim(0.5, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank())

a1 / c1

```

### Regression

```{r echo=FALSE}
all_engines_reg_min_med_fs <- all_engines_reg_min_med
all_engines_reg_min_med_fs <- all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Metric == 'rmse', ]
all_engines_reg_min_med_fs$Feature_selection <- ifelse(all_engines_reg_min_med_fs$Feature_selection != 'none', 'Yes', 'No')
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
d1 <- ggplot(data = all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Aggregation == 'Min', ],
            aes(x = Value, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title    = 'Minimal RMSE values (Magnified)',
       subtitle = 'for different regression tasks and preprocessing strategies',
       x        = 'RMSE',
       y        = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  coord_cartesian(xlim = c(0, 0.25)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.x  = element_blank(),
        axis.title.y  = element_blank())

e1 <- ggplot(data = all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Aggregation == 'Min', ],  
            aes(x = Value, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title    = 'Minimal RMSE values (All)',
       x        = 'RMSE',) +
  coord_cartesian(xlim = c(0, NA)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank(),
        axis.title.x   = element_blank())

f1 <- ggplot(data = all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Aggregation == 'Median', ],
            aes(x = Value, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title    = 'Median RMSE values (Magnified)',
       x        = 'RMSE',
       y        = 'Dataset',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  coord_cartesian(xlim = c(0, 0.25)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank())

g1 <- ggplot(data = all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Aggregation == 'Median', ], 
            aes(x = Value, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title    = 'Median RMSE values (All)',
       x        = 'RMSE',
       color = 'Are FS method used?',
       fill  = 'Are FS method used?') +
  coord_cartesian(xlim = c(0, NA)) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        plot.subtitle = element_blank(),
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank())

(d1 | e1) / (f1 | g1)

```

## Lack of Feature Selection

### Binary Classification

```{r echo=FALSE}
all_engines_bin_no_fs <- all_engines_bin_fs[all_engines_bin_fs$Feature_selection == 'No', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
h1 <- ggplot(data = all_engines_bin_no_fs, aes(x = Max, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, without FS, depending on removal strategy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.x  = element_blank(),
        axis.title.y  = element_blank())

i1 <- ggplot(data = all_engines_bin_no_fs, aes(x = Median, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

h1 / i1
```

### Multiclass classification

```{r echo=FALSE}
all_engines_mcl_no_fs <- all_engines_mcl_fs[all_engines_mcl_fs$Feature_selection == 'No', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
h1 <- ggplot(data = all_engines_mcl_no_fs, aes(x = Max, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks, without FS, depending on removal strategy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.x  = element_blank(),
        axis.title.y  = element_blank())

i1 <- ggplot(data = all_engines_mcl_no_fs, aes(x = Median, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

h1 / i1
```

### Regression

```{r echo=FALSE}
all_engines_reg_no_fs <- all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Feature_selection == 'No', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
l1 <- ggplot(data = all_engines_reg_no_fs[all_engines_reg_no_fs$Aggregation == 'Median', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median RMSE (Magnified)',
       subtitle = 'for regression tasks, without FS, depending on removal strategy',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

m1 <- ggplot(data = all_engines_reg_no_fs[all_engines_reg_no_fs$Aggregation == 'Median', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median RMSE',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank(),
        axis.text.y = element_blank(),
        plot.subtitle = element_blank())

n1 <- ggplot(data = all_engines_reg_no_fs[all_engines_reg_no_fs$Aggregation == 'Min', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Minimal (Magnified)',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

o1 <- ggplot(data = all_engines_reg_no_fs[all_engines_reg_no_fs$Aggregation == 'Min', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Minimal RMSE',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968","#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        axis.text.y   = element_blank(),
        plot.subtitle = element_blank())

(l1 | m1) / (n1 | o1)

```

## Feature selection only

### Binary classification

```{r echo=FALSE}
all_engines_bin_fs_only <- all_engines_bin_fs[all_engines_bin_fs$Feature_selection == 'Yes', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
p1 <- ggplot(data = all_engines_bin_fs_only, aes(x = Max, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_bin_baselines[all_engines_bin_baselines$Metric == 'accuracy', ], 
             size = 5, shape = 4, aes(x = Max, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, with FS, depending on removal strategy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

r1 <- ggplot(data = all_engines_bin_fs_only, aes(x = Median, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_bin_baselines[all_engines_bin_baselines$Metric == 'accuracy', ], 
             size = 5, shape = 4, aes(x = Median, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

p1 / r1
```

### Multiclass classification

```{r echo=FALSE}
all_engines_mcl_fs_only <- all_engines_mcl_fs[all_engines_mcl_fs$Feature_selection == 'Yes', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
p1 <- ggplot(data = all_engines_mcl_fs_only, aes(x = Max, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_mcl_baselines[all_engines_mcl_baselines$Metric == 'accuracy', ], 
             size = 5, shape = 4, aes(x = Max, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks, with FS, depending on removal strategy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

r1 <- ggplot(data = all_engines_mcl_fs_only, aes(x = Median, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_mcl_baselines[all_engines_mcl_baselines$Metric == 'accuracy', ], 
             size = 5, shape = 4, aes(x = Median, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

p1 / r1
```

### Regression

```{r echo=FALSE}
all_engines_reg_fs_only <- all_engines_reg_min_med_fs[all_engines_reg_min_med_fs$Feature_selection == 'Yes', ]
```

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
u1 <- ggplot(data = all_engines_reg_fs_only[all_engines_reg_fs_only$Aggregation == 'Median', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == 'rmse' & 
                                                            all_engines_reg_baselines_min_med$Aggregation == 'Median'), ], 
             size = 4, shape = 4, aes(x = Value, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Median RMSE (Magnified)',
       subtitle = 'with FS used for different binary classification tasks',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

v1 <- ggplot(data = all_engines_reg_fs_only[all_engines_reg_fs_only$Aggregation == 'Median', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == 'rmse' & 
                                                            all_engines_reg_baselines_min_med$Aggregation == 'Median'), ], 
             size = 4, shape = 4, aes(x = Value, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Median RMSE',
       subtitle = 'without FS used for different binary classification tasks',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank(),
        axis.text.y = element_blank())

w1 <- ggplot(data = all_engines_reg_fs_only[all_engines_reg_fs_only$Aggregation == 'Min', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == 'rmse' & 
                                                            all_engines_reg_baselines_min_med$Aggregation == 'Min'), ], 
             size = 4, shape = 4, aes(x = Value, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Minimal (Magnified)',
       subtitle = 'for different binary classification tasks',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

x1 <- ggplot(data = all_engines_reg_fs_only[all_engines_reg_fs_only$Aggregation == 'Min', ], 
             aes(x = Value, y = Dataset, color = factor(Removal), fill = factor(Removal))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = all_engines_reg_baselines_min_med[which(all_engines_reg_baselines_min_med$Metric == 'rmse' & 
                                                            all_engines_reg_baselines_min_med$Aggregation == 'Min'), ], 
             size = 4, shape = 4, aes(x = Value, y = Dataset), color = '#B1805B', fill = '#B1805B') +
  labs(title = 'Minimal RMSE',
       subtitle = 'for different binary classification tasks',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        axis.text.y  = element_blank(),
        plot.subtitle = element_blank())

(u1 | v1) / (w1 | x1)

```

## Feature Selection methods and performance

```{r}
all_engines_bin_fs_methods <- all_engines_bin[all_engines_bin$Feature_selection != 'none' & all_engines_bin$Metric == 'accuracy', ]
all_engines_mcl_fs_methods <- all_engines_mcl[all_engines_mcl$Feature_selection != 'none' & all_engines_mcl$Metric == 'accuracy', ]
all_engines_reg_fs_methods <- all_engines_reg[all_engines_reg$Feature_selection != 'none' & all_engines_reg$Metric == 'rmse', ]
```

### Binary classification

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
p1 <- ggplot(data = all_engines_bin_fs_methods, aes(x = Max, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, depending on feature selection method',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

r1 <- ggplot(data = all_engines_bin_fs_methods, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       subtitle = 'for binary classification tasks, depending on feature selection method',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.7, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank(),
        plot.subtitle = element_blank())

p1 / r1
```

### Multiclass classification

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
p1 <- ggplot(data = all_engines_mcl_fs_methods, aes(x = Max, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks, depending on feature selection method',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

r1 <- ggplot(data = all_engines_mcl_fs_methods, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       subtitle = 'for multiclass classification tasks, depending on feature selection method',
       x = 'Value',
       y = 'Dataset',
       color = 'Removal strategy',
       fill  = 'Removal strategy') +
  xlim(0.5, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank(),
        plot.subtitle = element_blank())

p1 / r1
```

### Regression

```{r fig.height=10, fig.width=18, echo=FALSE, warning=FALSE}
u1 <- ggplot(data = all_engines_reg_fs_methods, aes(x = Min, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median RMSE (Magnified)',
       subtitle = 'for regression tasks, depending on feature selection method',
       x = 'RMSE',
       y = 'Dataset') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())
v1 <- ggplot(data = all_engines_reg_fs_methods, aes(x = Min, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) +
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median RMSE',
       x = 'RMSE',
       y = 'Dataset') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank(),
        axis.text.y = element_blank())

w1 <- ggplot(data = all_engines_reg_fs_methods, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) +
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Minimal (Magnified)',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  xlim(0, 0.25) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

x1 <- ggplot(data = all_engines_reg_fs_methods, aes(x = Median, y = Dataset, color = factor(Feature_selection), fill = factor(Feature_selection))) +
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Minimal RMSE',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Feature selection method',
       fill  = 'Feature selection method') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        axis.text.y  = element_blank(),
        plot.subtitle = element_blank())

(u1 | v1) / (w1 | x1)

```


## Imputation impact

```{r echo=FALSE}
all_engines_bin_imp_no_fs <- all_engines_bin[all_engines_bin$Dataset %in% c('breast-w', 'credit-approval', "credit-g-mod", "phoneme-mod", "car-mod", "satimage-mod", "elevators-mod", "kin8nm-mod"), ]
all_engines_bin_imp_no_fs <- all_engines_bin_imp_no_fs[all_engines_bin_imp_no_fs$Feature_selection == 'none', ]

all_engines_mcl_imp_no_fs <- all_engines_mcl[all_engines_mcl$Dataset %in% c('breast-w', 'credit-approval', "credit-g-mod", "phoneme-mod", "car-mod", "satimage-mod", "elevators-mod", "kin8nm-mod"), ]
all_engines_mcl_imp_no_fs <- all_engines_mcl_imp_no_fs[all_engines_mcl_imp_no_fs$Feature_selection == 'none', ]

all_engines_reg_imp_no_fs <- all_engines_reg[all_engines_reg$Dataset %in% c('breast-w', 'credit-approval', "credit-g-mod", "phoneme-mod", "car-mod", "satimage-mod", "elevators-mod", "kin8nm-mod"), ]
all_engines_reg_imp_no_fs <- all_engines_reg_imp_no_fs[all_engines_reg_imp_no_fs$Feature_selection == 'none', ]
```


```{r fig.height=10, fig.width=20, echo=FALSE, warning=FALSE}
p1 <- ggplot(data = all_engines_bin_imp_no_fs, aes(x = Max, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, without FS,\ndepending on imputation method',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0.3, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

r1 <- ggplot(data = all_engines_bin_imp_no_fs, aes(x = Median, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0.3, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

s1 <- ggplot(data = all_engines_mcl_imp_no_fs, aes(x = Max, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0.3, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

t1 <- ggplot(data = all_engines_mcl_imp_no_fs, aes(x = Median, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median Accuracy',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0.3, 1) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "bottom",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

u1 <- ggplot(data = all_engines_reg_imp_no_fs[all_engines_reg_imp_no_fs$Metric == 'rmse', ], 
             aes(x = Min, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Min RMSE',
       subtitle = 'for regression tasks',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        axis.title.x = element_blank())

w1 <- ggplot(data = all_engines_reg_imp_no_fs[all_engines_reg_imp_no_fs$Metric == 'rmse', ], 
             aes(x = Median, y = Dataset, color = factor(Imputation), fill = factor(Imputation))) + 
  geom_boxplot(alpha = 0.5) +
  labs(title = 'Median RMSE',
       x = 'Value',
       y = 'Dataset',
       color = 'Imputation method',
       fill  = 'Imputation method') +
  xlim(0, NA) +
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(legend.position = "none",
        axis.title.y  = element_blank(),
        plot.subtitle = element_blank())

(p1 / r1) | (s1 / t1) | (u1 / w1)
```

## Engines comparison

```{r}
tree_engines <- validation_summary_table[validation_summary_table$Engine != 'all', ]

tree_engines_bin           <- tree_engines[tree_engines$Task_type == 'binary' & tree_engines$Metric == 'accuracy', ]
tree_engines_mcl           <- tree_engines[tree_engines$Task_type == 'multiclass' & tree_engines$Metric == 'accuracy', ]
tree_engines_reg           <- tree_engines[tree_engines$Task_type == 'regression' & tree_engines$Metric == 'rmse', ]

tree_engines_bin_baselines <- tree_engines_bin[which(tree_engines_bin$Removal =='removal_min' & 
                                                   tree_engines_bin$Imputation =='median-other' & 
                                                   tree_engines_bin$Feature_selection =='none'), ]
tree_engines_mcl_baselines <- tree_engines_mcl[which(tree_engines_mcl$Removal =='removal_min' & 
                                                   tree_engines_mcl$Imputation =='median-other' & 
                                                   tree_engines_mcl$Feature_selection =='none'), ]
tree_engines_reg_baselines <- tree_engines_reg[which(tree_engines_reg$Removal =='removal_min' & 
                                                   tree_engines_reg$Imputation =='median-other' & 
                                                   tree_engines_reg$Feature_selection =='none'), ]
```

### Classification tasks

```{r fig.height=12, fig.width=18, echo=FALSE, warning=FALSE}
o <- ggplot(data = tree_engines_bin, aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = tree_engines_bin_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) +
  labs(title = 'Max Accuracy',
       subtitle = 'for binary classification tasks, X-mark stands for baseline preprocessing',
       x = 'Value',
       y = 'Dataset',
       color = 'Engine',
       fill  = 'Engine') +
  xlim(0, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.title.x  = element_blank(),
        axis.title.y  = element_blank(),
        legend.position = "none")

p <- ggplot(data = tree_engines_mcl, aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = tree_engines_mcl_baselines, size = 4, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) +
  labs(title = 'Max Accuracy',
       subtitle = 'for multiclass classification tasks',
       x = 'Value',
       y = 'Dataset',
       color = 'Engine',
       fill  = 'Engine') +
  xlim(0, 1) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d",  "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.title.y  = element_blank(),
        legend.position = "bottom")
o / p
```

### Regression

```{r fig.height=12, fig.width=24, echo=FALSE, warning=FALSE, message=FALSE}
r <- ggplot(data = tree_engines_reg, aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) + 
  geom_boxplot(alpha = 0.5) +
  geom_point(data = tree_engines_reg_baselines, size = 6, shape = 4, 
             position = position_jitterdodge(), 
             aes(x = Max, y = Dataset, color = factor(Engine), fill = factor(Engine))) +
  labs(title = 'Max RMSE',
       subtitle = 'for regression tasks, with 5 different zooms, X-mark stands for baseline preprocessing',
       x = 'RMSE',
       y = 'Dataset',
       color = 'Engine',
       fill  = 'Engine') +
  xlim(0.0045, 0.007) + 
  paper_theme() +
  scale_color_manual(values = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  scale_fill_manual(values  = c("#afc968", "#74533d", "#7C843C", "#B1805B", '#D6E29C')) +
  theme(axis.title.y  = element_blank(),
        legend.position = "none")
s <- r +
  xlim(0.09, 0.14) +
  theme(axis.title.y  = element_blank(),
        axis.text.y  = element_blank(),
        plot.subtitle = element_blank(),
        plot.title = element_blank())

t <- s +
  xlim(0.2, 0.26) +
  theme(legend.position = "bottom")

u <- t +
  xlim(0.95, 4.5) +
  theme(legend.position = "none") 

w <- u +
  xlim(6, 17) 

r | s | t | u | w
```
